CornerText: Read DataProbe toggled corners & Implement Widget
A host of changes were made to integrate the logic of CornerText with
the Widget of the loadable module. First, settings from DataProbe are
read. Then, CornerText copies the same UI as DataProbe but connects the
signals from CornerText's widget to the CornerText logic. Therefore,
once DataProbe settings are read, the behavior of CornerText is
independent but identical in function.

- CornerText right now omits the implementation of individual slice node
  corners and instead reads the settings from DataProbe to globally
  toggle certain corners.  The way this works is 1) Nothing about
  toggled corners within Slice Nodes is serialized, and 2) Whenever the
  layout changes the Widget updates (which is also updated when a
  checkbox is selected), which calls
  `vtkMRMLCornerTextLogic::ToggleLocationGlobally` to loop through all existing
  Slice Nodes in the scene and toggle the flags for all the slice nodes
  given the global settings.

- Within the DM, overloaded `OnMRMLDisplayableNodeModifiedEvent` and
  added observations to modifications of the corner text logic to
  increase the number of times the DM will propagate updates to the
  slice node.

- ToggleLocationGlobally was moved to the Widget, as it iterates over
  the slice nodes which are a part of the scene. CornerTextLogic should
  be agnostic to the scene, and should be supplied the slice node. In
  fact, CornerTextLogic->GetMRMLScene() returns null. While this may be
  undesirable for now, the solution still makes sense to move the
  ToggleLocationGlobally to either the Widget, Layout, or DM (anything
  which will inherently be tied to a collection of slice nodes).

- Refactored much of the logic in the DM's
  `UpdateCornerAnnotationsFromSliceNode` so that 1) it actually sets the
  text to an empty string when a corner is disabled, and 2) does not
  update the corner annotations as many times. TODO: I could also set it
  from a vector with empty strings in the indices, but this may not
  bring much of an improvement...

Overall, with a few shortcomings, CornerText is implemented as an
alternative DataProbe. We should still read settings and set the FooBar
widget parameters, for aesthetic reasons (though there isn't anything
broken with how the FooBar widget propagates updates to the
CornerTextLogic). We should also observe the text node in the DM and
make updates then. There are also several efficiency considerations, and
Slicer flat-out crashes when creating new slice views (in addition to
the existing ones).