ENH: Add DICOM corner annotations provider (#7184)
Added a DICOM-specific property value provider to the corner annotation
system, enabling the display of DICOM metadata such as patient and study
details. The provider replicates the previous behavior of DataProbe,
using Python for database queries. Currently, it uses a high-level
wrapper to static functions defined in
`DICOMAnnotationPropertyValueProvider.py`, but this should be optimized
if the API logic is not adapted further. The purpose of these changes
are to demonstrate the extensibility of the refactored annotation
infrastructure and ensure compatibility with existing annotation
functionality.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>